+++ grafanalib/tests/test_grafanalib.py
@@ -61,6 +61,30 @@ def test_auto_id():
     ).auto_panel_ids()
     assert dashboard.rows[0].panels[0].id == 1
 
+    dashboard = G.Dashboard(
+        title="Test dashboard",
+        panels=[
+            G.RowPanel(gridPos=G.GridPos(h=1, w=24, x=0, y=8)),
+            G.Graph(
+                title="CPU Usage by Namespace (rate[5m])",
+                dataSource="My data source",
+                targets=[
+                    G.Target(
+                        expr='whatever',
+                        legendFormat='{{namespace}}',
+                        refId='A',
+                    ),
+                ],
+                yAxes=G.YAxes(
+                    G.YAxis(format=G.SHORT_FORMAT, label="CPU seconds"),
+                    G.YAxis(format=G.SHORT_FORMAT),
+                ),
+                gridPos=G.GridPos(h=1, w=24, x=0, y=8)
+            )
+        ],
+    ).auto_panel_ids()
+    assert dashboard.panels[0].id == 1
+
 
 def test_auto_refids_preserves_provided_ids():
     """
@@ -94,6 +118,38 @@ def test_auto_refids_preserves_provided_ids():
     assert dashboard.rows[0].panels[0].targets[0].refId == 'A'
     assert dashboard.rows[0].panels[0].targets[1].refId == 'Q'
     assert dashboard.rows[0].panels[0].targets[2].refId == 'B'
+
+    dashboard = G.Dashboard(
+        title="Test dashboard",
+        panels=[
+            G.Graph(
+                title="CPU Usage by Namespace (rate[5m])",
+                dataSource="My data source",
+                targets=[
+                    G.Target(
+                        expr='whatever #Q',
+                        legendFormat='{{namespace}}',
+                    ),
+                    G.Target(
+                        expr='hidden whatever',
+                        legendFormat='{{namespace}}',
+                        refId='Q',
+                    ),
+                    G.Target(
+                        expr='another target'
+                    ),
+                ],
+                yAxes=G.YAxes(
+                    G.YAxis(format=G.SHORT_FORMAT, label="CPU seconds"),
+                    G.YAxis(format=G.SHORT_FORMAT),
+                ),
+                gridPos=G.GridPos(h=1, w=24, x=0, y=8)
+            ).auto_ref_ids()
+        ],
+    ).auto_panel_ids()
+    assert dashboard.panels[0].targets[0].refId == 'A'
+    assert dashboard.panels[0].targets[1].refId == 'Q'
+    assert dashboard.panels[0].targets[2].refId == 'B'
 
 
 def test_auto_refids():
@@ -132,3 +188,16 @@ def test_row_show_title():
     row = G.Row(title='My title', showTitle=False).to_json_data()
     assert row['title'] == 'My title'
     assert not row['showTitle']
+
+
+def test_row_panel_show_title():
+    row = G.RowPanel().to_json_data()
+    assert row['title'] == ''
+    assert row['panels'] == []
+
+    row = G.RowPanel(title='My title').to_json_data()
+    assert row['title'] == 'My title'
+
+    row = G.RowPanel(title='My title', panels=['a', 'b']).to_json_data()
+    assert row['title'] == 'My title'
+    assert row['panels'][0] == 'a'