+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Changelog
 * Added Logs panel (https://grafana.com/docs/grafana/latest/panels/visualizations/logs-panel/)
 * Added Cloudwatch metrics datasource (https://grafana.com/docs/grafana/latest/datasources/cloudwatch/)
 * Added option to hide dashboard time picker
+* Added Notification for Alert
+* Added alertRuleTags field to the graph panel
+* Added support for using gridPos for dashboard panels
 
 Changes
 -------